+++ client/modules/_helpers.js
@@ -1,0 +1,6 @@
+export const htmlEscapeStr = str => {
+  return str.replace(/&/g, '&amp;')
+    .replace(/>/g, '&gt;')
+    .replace(/</g, '&lt;')
+    .replace(/"/g, '&quot;');
+};