+++ server/modules/postgres_store.js
@@ -7,8 +7,8 @@
  */
 
 import postgres from 'pg';
-import winston from 'winston';
-import Config from '../../config';
+import Config from '../config/config';
+import { privacyOptions } from '../../shared/config/constants';
 import { postgresTimestamp } from './_helpers';
 
 class PostgresStore {
@@ -19,8 +19,8 @@ class PostgresStore {
 
   insert(key, data, options = {}) {
     return new Promise(async (resolve, reject) => {
-      const queryString = 'INSERT INTO entries (key, text, public, name, title, expiration) VALUES ($1, $2, $3, $4, $5, $6)';
-      const queryKeys = [key, data.text, data.public, data.name, data.title, data.expiration];
+      const queryString = 'INSERT INTO entries (key, text, privacyOption, name, title, expiration) VALUES ($1, $2, $3, $4, $5, $6)';
+      const queryKeys = [key, data.text, data.privacyOption, data.name, data.title, data.expiration];
       await this.query(queryString, queryKeys);
       resolve(true);
       reject(false);
@@ -41,8 +41,8 @@ class PostgresStore {
   getList() {
     return new Promise(async (resolve, reject) => {
       const now = postgresTimestamp();
-      const queryString = 'SELECT * from entries where public = $1 and (expiration IS NULL or expiration > $2)';
-      const queryKeys = [true, now];
+      const queryString = 'SELECT * from entries where privacyOption = $1 and (expiration IS NULL or expiration > $2)';
+      const queryKeys = [privacyOptions.public, now];
       const { rows } = await this.query(queryString, queryKeys);
       resolve(rows.length ? rows : false);
       reject(false);